Move parameter handling to separate module
The parameter handling code with all of its tests is actually longer
than the main code. This splits it off into a separate module, mostly to
keep the tests separate from any other tests we add in the future.
